edupulseai/requirements.txt
pandas
numpy
scikit-learn
streamlit
matplotlib
plotly
pydantic
pyyaml
joblib
pytest
requests
